fix internal server error if user doesn't exist

+++ app/blueprints/users.py
@@ -25,12 +25,16 @@ def read_all():
 @users.route('/<user_id>', methods=['GET'])
 def read_one(user_id):
     user = User.query.get(user_id)
+    if user is None:
+        return '', 404
     return jsonify(user.json())
 
 
 @users.route('/<user_id>', methods=['PUT', 'PATCH'])
 def update(user_id):
     user = User.query.get(user_id)
+    if user is None:
+        return '', 404
     if request.json.get('first_name'):
         user.first_name = request.json['first_name']
     if request.json.get('last_name'):
@@ -44,6 +48,8 @@ def update(user_id):
 @users.route('/<user_id>', methods=['DELETE'])
 def delete(user_id):
     user = User.query.get(user_id)
+    if user is None:
+        return '', 404
     db.session.delete(user)
     db.session.commit()
     return '', 204
